Pull SacnController out of SacnManager.
